App working now (except edge & ie)
