+++ src/Components/Second Page/Bills/AddBill/style.css
@@ -5,56 +5,3 @@
 input[type='number'] {
   -moz-appearance: textfield;
 }
-
-input::-webkit-outer-spin-button,
-input::-webkit-inner-spin-button {
-  -webkit-appearance: none;
-}
-
-.toggle {
-  -webkit-appearance: none;
-  -moz-appearance: none;
-  appearance: none;
-  width: 56px;
-  height: 26px;
-  display: inline-block;
-  position: relative;
-  border-radius: 50px;
-  overflow: hidden;
-  border: none;
-  cursor: pointer;
-  background-color: #707070;
-  transition: ease 0.3s;
-  margin: -6px auto;
-}
-
-.toggle:before {
-  content: '';
-  display: block;
-  position: absolute;
-  z-index: 2;
-  width: 22px;
-  height: 22px;
-  background: #fff;
-  left: 2px;
-  top: 2px;
-  border-radius: 50%;
-  font: 20px/28px Helvetica;
-  text-transform: uppercase;
-  font-weight: bold;
-  text-indent: -22px;
-  word-spacing: 37px;
-  color: #fff;
-  text-shadow: -1px -1px rgba(0, 0, 0, 0.15);
-  white-space: nowrap;
-  box-shadow: 0 1px 2px rgba(0, 0, 0, 0.2);
-  transition: ease-in 0.3s;
-}
-
-.toggle:checked {
-  background-color: #4cd964;
-}
-
-.toggle:checked:before {
-  left: 32px;
-}
